move user fixtures to conftest.py
